% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoreg.R
\name{isoreg_Best_1990}
\alias{isoreg_Best_1990}
\title{Isotonic / Monotone Regression}
\usage{
isoreg_Best_1990(
  x,
  y = NULL,
  decreasing = FALSE,
  implementation = c("cpp", "R")
)
}
\arguments{
\item{x}{abscissa values (\code{x}) if \code{y} is supplied else target values.}

\item{y}{numeric vector target values.}

\item{decreasing}{logical indicating if a monotonic decreasing should be fit (instead of monotonic increasing)}

\item{implementation}{a character string specifying if the fast C++ or an equivalent R implementation should be used.}
}
\value{
numeric vector of fitted values (\code{yf}).
}
\description{
Compute the isotonic (monotonely increasing nonparametric) least squares regression which is piecewise constant.
}
\details{
If abscissa values are supplied the "primary" tie method is performed \eqn{(y_i<y_j \imply yf_i<yf_j)}.
}
\examples{
y <- c(1, 3, 2, 4)
isoreg_Best_1990(y)

x <- c(1, 0, 3, 2)
y <- c(3, 1, 4, 2)
isoreg_Best_1990(x, y)
}
