% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rank_in_cpp}
\alias{rank_in_cpp}
\title{Compute rank of Q in X}
\usage{
rank_in_cpp(X, Q, nsw = 400L)
}
\arguments{
\item{X}{Matrix of target values}

\item{Q}{Matrix of query values}

\item{nsw}{integer to tune hybrid algorithm}
}
\value{
vector of rank of each row of Q in X
}
\description{
Compute rank of Q in X
}
