% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isoreg_Best_1990_cpp}
\alias{isoreg_Best_1990_cpp}
\title{Isotonic regression after Best & Chakravarti (1990)}
\usage{
isoreg_Best_1990_cpp(y)
}
\arguments{
\item{y}{target values.}
}
\value{
fitted values.
}
\description{
C++ implementation of Algorithm A from Best & Chakravarti (1990).
Performs isotonic regression with a run time complexity of \eqn{\mathcal{O}(n)}
}
\details{
Best, M.J. & Chakravarti, N. Mathematical Programming (1990) 47: 425.
\href{http://doi.org/10.1007/BF01580873}{doi:10.1007/BF01580873}
}
\examples{
y <- c(1, 3, 2, 4)
isoreg_Best_1990_cpp(y)
}
\author{
Jan Gleixner
}
